Fix Kingdom collection
