+++ static/lang/kingmaker-en.json
@@ -175,8 +175,9 @@
             "CollectDialog": {
                 "Title": "Collect Resources",
                 "Header": "Upkeep: Resource Collection",
-                "Description": "Roll {dice} resource dice and generate the following?",
-                "Disclaimer": "This will replace your current resource points (which should have been converted into experience points). Excess commodities will be lost. Only the game master can collect resources."
+                "Description": "Roll {dice} resource dice and also generate the following?",
+                "Disclaimer": "This will replace your current resource points (which should have been converted into experience points). Excess commodities will be lost. Only the game master can collect resources.",
+                "NewResourcePoints": "New Resource Points"
             },
             "Commodity": {
                 "food": "Food",
