+++ static/templates/actors/party/kingdom/collection.hbs
@@ -2,14 +2,16 @@
     <header class="card-header flexrow">
         <h3>{{localize "PF2E.Kingmaker.Kingdom.CollectDialog.Header"}}</h3>
     </header>
-    <p>{{localize "PF2E.Kingmaker.Kingdom.CollectDialog.Description" dice=formula}}</p>
-    <ul>
-    {{#each commodities as |value slug|}}
-        <li>{{value}} {{localize (concat "PF2E.Kingmaker.Kingdom.Commodity." slug)}}</li>
-    {{/each}}
-    </ul>
-    <p>{{localize "PF2E.Kingmaker.Kingdom.CollectDialog.Disclaimer"}}</p>
-    <button type="button" data-action="kingdom-collect" data-visibility="owner">
+    <section class="card-content">
+        <p>{{localize "PF2E.Kingmaker.Kingdom.CollectDialog.Description" dice=formula}}</p>
+        <ul>
+        {{#each commodities as |value slug|}}
+            <li>{{value}} {{localize (concat "PF2E.Kingmaker.Kingdom.Commodity." slug)}}</li>
+        {{/each}}
+        </ul>
+        <p>{{localize "PF2E.Kingmaker.Kingdom.CollectDialog.Disclaimer"}}</p>
+    </section>
+    <button type="button" data-action="kingdomCollect" data-visibility="owner">
         {{localize "PF2E.Kingmaker.Confirm"}}
     </button>
 </div>
